Added in house middleware support
Created a middleware for converting files given specific parsers.
This will require much better documentation. Currently the only
support is for less and the parser is uncommented.
Will need to comment this parser, iron out bugs in the middleware
and create a parser for coffee script to finish this version minor.